Conflicts:
	nbactions.xml
	pom.xml
	src/main/java/org/uv/tpcs_practica01/ConexionDB.java
	src/main/java/org/uv/tpcs_practica01/DAOEmpleado.java
	src/main/java/org/uv/tpcs_practica01/Principal.form
	src/main/java/org/uv/tpcs_practica01/Principal.java
	src/main/java/org/uv/tpcs_practica01/TPCS_Practica01.java

+++ src/main/java/org/uv/tpcs_practica01/Empleado.java
@@ -6,7 +6,7 @@ public class Empleado {
         return clave;
     }
 
-    public void setClave(long clave) {
+    public void setClave(int clave) {
         this.clave = clave;
     }
